Fetch records asynchronously, preemptively
BufferedGetter now uses a concurrent queue and an ExecutorService with which it
will execute tasks that will fetch additional records and append them to the end
of the queue. Tasks for fetching record lists are fired preemptively and are
limited to running once every 200ms (to comply with the service limit of 5
reads/second).

Uses a method reference, meaning a dependency on Java 1.8.